# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistema Pesquisas — API Multi-Tenant",
        "description": "Backend REST API multiempresa com RBAC. Use POST /api/v1/auth/login para obter o JWT e clique em Authorize inserindo: Bearer {token}. Super Admins podem usar o header X-Empresa-Id para escopar requisições a uma empresa.",
        "contact": {
            "email": "admin@sistema.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/perfil": {
            "get": {
                "tags": [
                    "Perfil"
                ],
                "summary": "Ver próprio perfil",
                "operationId": "466a966b9931420cb710b3ac1b2ff036",
                "responses": {
                    "200": {
                        "description": "Perfil do usuário",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sucesso": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "dados": {
                                            "$ref": "#/components/schemas/Usuario"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Perfil"
                ],
                "summary": "Atualizar próprio perfil",
                "operationId": "bfd265eb92abfe391d1f8c1c309e805d",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "João Silva"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "telefone": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Perfil atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sucesso": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "dados": {
                                            "$ref": "#/components/schemas/Usuario"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/perfil/trocar-senha": {
            "post": {
                "tags": [
                    "Perfil"
                ],
                "summary": "Trocar própria senha",
                "operationId": "5d3d4ce58e85242ca98989e921ab7145",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "senha_atual",
                                    "nova_senha",
                                    "confirmacao"
                                ],
                                "properties": {
                                    "senha_atual": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "nova_senha": {
                                        "type": "string",
                                        "format": "password",
                                        "minLength": 8
                                    },
                                    "confirmacao": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Senha alterada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sucesso": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "mensagem": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/perfil/foto": {
            "post": {
                "tags": [
                    "Perfil"
                ],
                "summary": "Upload da foto de perfil",
                "operationId": "61d86f5f3a9be9d573c93d98d9924ca1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "foto": {
                                        "description": "JPG/PNG/GIF máx 2MB",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Foto atualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sucesso": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/perfil/permissoes": {
            "get": {
                "tags": [
                    "Perfil"
                ],
                "summary": "Listar permissões do usuário",
                "description": "Todas as permissões consolidadas dos roles do usuário.",
                "operationId": "2d1cc8b5d0ffadbb09d1429e975a33d3",
                "responses": {
                    "200": {
                        "description": "Lista de permissões",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sucesso": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "dados": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Permission"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Permission": {
                "description": "Permissão do sistema",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nome": {
                        "type": "string",
                        "example": "Listar Empresas"
                    },
                    "slug": {
                        "type": "string",
                        "example": "empresa.listar"
                    },
                    "modulo": {
                        "type": "string",
                        "example": "empresa"
                    },
                    "descricao": {
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Usuario": {
                "description": "Dados de um usuário",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 2
                    },
                    "empresa_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "filial_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "nome": {
                        "type": "string",
                        "example": "João Silva"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "joao@empresa.com"
                    },
                    "foto": {
                        "type": "string",
                        "nullable": true
                    },
                    "telefone": {
                        "type": "string",
                        "nullable": true
                    },
                    "tipo": {
                        "type": "string",
                        "example": "usuario",
                        "enum": [
                            "superadmin",
                            "admin",
                            "gerente",
                            "usuario"
                        ]
                    },
                    "status": {
                        "type": "string",
                        "example": "ativo",
                        "enum": [
                            "ativo",
                            "inativo",
                            "bloqueado"
                        ]
                    },
                    "ultimo_login": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "primeiro_acesso": {
                        "type": "boolean",
                        "example": false
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token JWT obtido via POST /api/v1/auth/login. Insira: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}